adjust options and instantiation for better external use
